+++ tests/integ/__init__.py
@@ -16,11 +16,14 @@ import logging
 import os
 import sys
 
+import boto3
+
 DATA_DIR = os.path.join(os.path.dirname(__file__), '..', 'data')
 TRAINING_DEFAULT_TIMEOUT_MINUTES = 20
 TUNING_DEFAULT_TIMEOUT_MINUTES = 20
 TRANSFORM_DEFAULT_TIMEOUT_MINUTES = 20
 PYTHON_VERSION = 'py' + str(sys.version_info.major)
+REGION = boto3.session.Session().region_name
 
 logging.getLogger('boto3').setLevel(logging.INFO)
 logging.getLogger('botocore').setLevel(logging.INFO)