+++ tests/integ/test_tf_keras.py
@@ -18,12 +18,14 @@ import numpy as np
 import pytest
 
 from sagemaker.tensorflow import TensorFlow
-from tests.integ import DATA_DIR, PYTHON_VERSION
+from tests.integ import DATA_DIR, PYTHON_VERSION, REGION
 from tests.integ.timeout import timeout_and_delete_endpoint_by_name, timeout
 
 
 @pytest.mark.continuous_testing
 @pytest.mark.skipif(PYTHON_VERSION != 'py2', reason="TensorFlow image supports only python 2.")
+@pytest.mark.skipif(REGION in ['us-west-1', 'eu-west-2', 'ca-central-1'],
+                    reason='No ml.p2.xlarge supported in these regions')
 def test_keras(sagemaker_session, tf_full_version):
     script_path = os.path.join(DATA_DIR, 'cifar_10', 'source')
     dataset_path = os.path.join(DATA_DIR, 'cifar_10', 'data')